Bugfix for double strike
Two strikes in a row were not properly accounted for. The previous code
would only consider adding the points from both rolls of the current
frame as bonus if the previous frame was a strike. But when the current
frame is a strike as well, then also the points from the next roll in
the following frame need to be added as a bonus to the previous frame.

I noticed that after added a max score test (300 points).